refactor: centralize id generation
