+++ controllers/clientesController.js
@@ -1,5 +1,6 @@
 const supabase = require('../supabaseClient');
 const { assertSupabase } = require('../supabaseClient');
+const { gerarIdInterno, gerarIdUnico } = require('../utils/idGenerator');
 
 function sanitizeCpf(s = '') {
   return (s.match(/\d/g) || []).join('');
@@ -162,27 +163,6 @@ exports.remove = async (req, res) => {
   }
 };
 
-function gerarIdInterno() {
-  const digits = '23456789';
-  let out = 'C';
-  for (let i = 0; i < 7; i++) {
-    out += digits[Math.floor(Math.random() * digits.length)];
-  }
-  return out;
-}
-
-async function gerarIdUnico() {
-  while (true) {
-    const id = gerarIdInterno();
-    const { data } = await supabase
-      .from('clientes')
-      .select('id_interno')
-      .eq('id_interno', id)
-      .maybeSingle();
-    if (!data) return id;
-  }
-}
-
 exports.generateIds = async (req, res) => {
   try {
     if (!assertSupabase(res)) return;
@@ -194,7 +174,7 @@ exports.generateIds = async (req, res) => {
 
     let updated = 0;
     for (const cli of clientes || []) {
-      const novo = await gerarIdUnico();
+      const novo = await gerarIdUnico(supabase);
       const { error: upErr } = await supabase
         .from('clientes')
         .update({ id_interno: novo })
